refactor: refactor `useHightlight` to pure utility functions
